Add files via upload
Notify.png shows the error description.

Created a function instead of button click. and call this function twice in page load. Both function call is executing and log for those is created in browsers. But the notification shadow is visible at top corner of the browser window. Please check the 'Notify.png' file.